Make `new` have a lazy dependency for `redis`
We will currently create a Redis instance. We may want to change it to a
Redis::Fast, and extend it to it that it will try use `Redis::Fast`, and
fall back to `Redis` if not available.

Not sure what should be done if Redis wouldn't be available, it is
currently a CPAN required dependency, it should be advisory instead.